Cloud/functions/requirements.txt
functions-framework
google-cloud-bigquery
requests